+++ neobisss/src/Week4/src/main/java/org/example/onlineshop/entity/Product.java
@@ -7,13 +7,15 @@ import jakarta.persistence.Id;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.*;
+import java.math.BigDecimal;
+
 @Data
 @Entity
 public class Product {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Long id;
     @NotBlank
     private String productName;
     @NotNull
-    private double price;
+    private BigDecimal price;
 }
